Allow no intervening label between call and goto.
As discussed here:
https://groups.google.com/d/topic/tlaplus/6M1oFOtN-5k/discussion

If a call is followed directly by a goto, we can simply use the goto's
target as the return label for the call, thus negating the requirement
for the call to be followed by a label. This helps eliminate dummy
labels/states for certain cases.

The implementation is simply an adaptation of the existing machinery for
a call followed directly by return.

[Feature][PlusCal][Changelog]